Extract parsing logic
